scripts/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile ./scripts/requirements.in
#
pyjwt==2.3.0
    # via -r ./scripts/requirements.in
python-dotenv==0.19.2
    # via -r ./scripts/requirements.in
